Polyfill some hyperbolic functions in lambdify for numpy quaddtype

+++ tests/test_qoi_abs_pw.py
@@ -34,7 +34,7 @@ def check_unit(qoi: str):
 
 
 def check_circle(qoi: str):
-    check_all_codecs(np.linspace(-np.pi * 2, np.pi * 2, 100), qoi)
+    check_all_codecs(np.linspace(-np.pi * 2, np.pi * 2, 100, dtype=np.int64), qoi)
 
 
 def check_arange(qoi: str):
